requirements.txt
pytest
flask
psycopg[binary]
requests
python-dotenv
